+++ tools/vaspberry_transport.py
@@ -34,7 +34,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 KB_EV_PER_K = 8.617333262145e-5
 TWO_PI = 2.0 * math.pi